Fix issue #3 - Change position and matchCount from uint8 to uint16 to allow parsing data longer than 255 chars

+++ src/MicroXPath_P.h
@@ -81,8 +81,8 @@ class MicroXPath_P {
     uint8_t checksum;
     #endif
     uint8_t level;
-    uint8_t position;
-    uint8_t matchCount;
+    uint16_t position;
+    uint16_t matchCount;
     uint8_t matchLevel;
 
     bool find(char charToParse /* , const char *attributeName, const char *attributeValue */);
